feat(utxo-staking): rename and enhance babylon scripts
Rename babylon-params.ts to babylon-sync.ts and add functionality to sync
BTC delegation responses from Babylon API. Add better error handling and
fixture data for testing.

Issue: BTC-2143
